Working search for single course dept+num
Use JSON module for parsing.py instead of trying to do it by hand. Added courses.json to be accessible resource in manifest.json. Added drawPopup function to create a simple center-screen popup that will display info. DrawBubble draws the bubble that the website already uses sparsely. Now we create a link and insert a bubble on clicking that link (content.js)